Ubuntu/WSL port resurrection (updated to llvm-project 22.1.0 in that env.)
- Disable:

```
   -Werror
   -Wno-error=unused-but-set-variable
   -Wno-comment
   -Wno-overloaded-virtual
   -Wno-unused-parameter
```

for g++ compilation (g++-13/14 w/ -Werror doesn't get along with 22.1.0)

- Only use: -fdiagnostics-all-candidates w/ gcc 14+

- README: note ubuntu dwarfdump dependency.

- driver:  assemble with PIC options, passing -DUBUNTU_PIC_REQUIRED=1 to pick that option.

- main addressing is pickier in g++.  pass it using a function pointer type, until ready to finally cast to void*

- Restore the 'java -jar' command line for building the grammar from antlr-4.10-complete.jar (when detected.)

- __silly_print: adjust the fprintf to avoid compiler message.

- debug/function-call.silly: make the test less brittle.  check for external, not void AT_type.

+++ src/runtime/Silly_runtime.cpp
@@ -54,7 +54,7 @@ extern "C"
                     break;
 
                 default:
-                    fprintf( stderr, "Unknown PrintKind %u, for argument %d of %d\n", arg.kind, i, num_args );
+                    fprintf( stderr, "Unknown PrintKind %u, for argument %d of %d\n", (unsigned)arg.kind, i, num_args );
                     abort();
             }
         }
